requirements.txt
aiogram==3.6.0
python-dotenv==1.0.1
asyncpg==0.29.0
matplotlib==3.9.0
